gatewayEMS/pyproject.toml
[project]
name = "gatewayems"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pydantic-settings>=2.11.0",
    "pymodbus>=3.11.2",
    "pyserial>=3.5",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "watchdog>=6.0.0",
]
